Syncing updates and switching drivers

+++ capstone/src/main/java/com/techelevator/view/Candy.java
@@ -1,9 +1,11 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Candy extends Item {
 
 
-    public Candy(String slotID, String name, double price, String message) {
+    public Candy(String slotID, String name, BigDecimal price, String message) {
         super(slotID, name, price, message);
