+++ capstone/src/main/java/com/techelevator/view/Chip.java
@@ -1,9 +1,11 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Chip extends Item {
 
 
-    public Chip(String slotID, String name, double price, String message) {
+    public Chip(String slotID, String name, BigDecimal price, String message) {
         super(slotID, name, price, message);
     }
 }
