+++ capstone/src/main/java/com/techelevator/view/Drink.java
@@ -1,9 +1,11 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Drink extends Item{
 
 
-    public Drink(String slotID, String name, double price, String message) {
+    public Drink(String slotID, String name, BigDecimal price, String message) {
         super(slotID, name, price, message);
     }
 }
