+++ capstone/src/main/java/com/techelevator/view/Gum.java
@@ -1,9 +1,11 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Gum extends Item {
 
 
-    public Gum(String slotID, String name, double price, String message) {
+    public Gum(String slotID, String name, BigDecimal price, String message) {
         super(slotID, name, price, message);
     }
 }
